Move api and map view controllers into a controllers package.
