+++ tethysapp/temp_precip_trends/tests/integrated_tests/api.py
@@ -3,7 +3,7 @@ import json
 
 from rest_framework.test import APIRequestFactory
 from rest_framework.test import force_authenticate
-from tethysapp.temp_precip_trends import api
+from tethysapp.temp_precip_trends.controllers import api
 
 
 class ApiTests(TestCase):
@@ -14,8 +14,8 @@ class ApiTests(TestCase):
     def tearDown(self):
         pass
 
-    @mock.patch('tethysapp.temp_precip_trends.api.param_check')
-    @mock.patch('tethysapp.temp_precip_trends.api.get_data')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.param_check')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.get_data')
     def test_get_stats_temp_and_total_precip(self, mock_get_data, mock_param_check):
         mock_request_min = self.factory.get('/get-min-temp/')
         mock_request_mean = self.factory.get('/get-mean-temp/')
@@ -42,8 +42,8 @@ class ApiTests(TestCase):
         self.assertEqual(expected_result, json.loads(ret_max.content))
         self.assertEqual(expected_result, json.loads(ret_tp.content))
 
-    @mock.patch('tethysapp.temp_precip_trends.api.param_check')
-    @mock.patch('tethysapp.temp_precip_trends.api.get_data')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.param_check')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.get_data')
     def test_get_stats_temp_and_total_precip_exceptions(self, mock_get_data, mock_param_check):
         mock_request_min = self.factory.get('/get-min-temp/')
         mock_request_mean = self.factory.get('/get-mean-temp/')
@@ -70,8 +70,8 @@ class ApiTests(TestCase):
         self.assertEqual(expected_result, json.loads(ret_max.content))
         self.assertEqual(expected_result, json.loads(ret_tp.content))
 
-    @mock.patch('tethysapp.temp_precip_trends.api.param_check')
-    @mock.patch('tethysapp.temp_precip_trends.api.get_cum_precip_data')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.param_check')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.get_cum_precip_data')
     def test_get_cum_precip(self, mock_get_cpd, mock_param_check):
         mock_request = self.factory.get('/get-cum-precip/')
 
@@ -86,8 +86,8 @@ class ApiTests(TestCase):
 
         self.assertEqual(expected_result, json.loads(ret.content))
 
-    @mock.patch('tethysapp.temp_precip_trends.api.param_check')
-    @mock.patch('tethysapp.temp_precip_trends.api.get_cum_precip_data')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.param_check')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.get_cum_precip_data')
     def test_get_cum_precip_exceptions(self, mock_get_cpd, mock_param_check):
         mock_request = self.factory.get('/get-cum-precip/')
 
@@ -102,11 +102,11 @@ class ApiTests(TestCase):
 
         self.assertEqual(expected_result, json.loads(ret.content))
 
-    @mock.patch('tethysapp.temp_precip_trends.api.relativedelta')
-    @mock.patch('tethysapp.temp_precip_trends.api.datetime')
-    @mock.patch('tethysapp.temp_precip_trends.api.overlap_ts')
-    @mock.patch('tethysapp.temp_precip_trends.api.param_check')
-    @mock.patch('tethysapp.temp_precip_trends.api.get_data')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.relativedelta')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.datetime')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.overlap_ts')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.param_check')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.get_data')
     def test_get_proj_mean_temp(self, mock_get_data, mock_param_check, mock_overlap_ts, _, __):
         mock_request = self.factory.get('/get-proj-mean_temp/', {'end_time': 'test'})
 
@@ -122,11 +122,11 @@ class ApiTests(TestCase):
 
         self.assertEqual(expected_result, json.loads(ret.content))
 
-    @mock.patch('tethysapp.temp_precip_trends.api.relativedelta')
-    @mock.patch('tethysapp.temp_precip_trends.api.datetime')
-    @mock.patch('tethysapp.temp_precip_trends.api.overlap_ts')
-    @mock.patch('tethysapp.temp_precip_trends.api.param_check')
-    @mock.patch('tethysapp.temp_precip_trends.api.get_data')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.relativedelta')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.datetime')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.overlap_ts')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.param_check')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.get_data')
     def test_get_proj_mean_temp_exceptions(self, mock_get_data, mock_param_check, mock_overlap_ts, _, __):
         mock_request = self.factory.get('/get-proj-mean_temp/', {'end_time': 'test'})
 
@@ -142,11 +142,11 @@ class ApiTests(TestCase):
 
         self.assertEqual(expected_result, json.loads(ret.content))
 
-    @mock.patch('tethysapp.temp_precip_trends.api.relativedelta')
-    @mock.patch('tethysapp.temp_precip_trends.api.datetime')
-    @mock.patch('tethysapp.temp_precip_trends.api.overlap_ts')
-    @mock.patch('tethysapp.temp_precip_trends.api.param_check')
-    @mock.patch('tethysapp.temp_precip_trends.api.get_cum_precip_data')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.relativedelta')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.datetime')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.overlap_ts')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.param_check')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.get_cum_precip_data')
     def test_get_proj_cum_precip(self, mock_get_cpd, mock_param_check, mock_overlap_ts, _, __):
         mock_request = self.factory.get('/get-proj-cum-precip/', {'end_time': 'test'})
 
@@ -162,12 +162,12 @@ class ApiTests(TestCase):
 
         self.assertEqual(expected_result, json.loads(ret.content))
 
-    @mock.patch('tethysapp.temp_precip_trends.api.log')
-    @mock.patch('tethysapp.temp_precip_trends.api.relativedelta')
-    @mock.patch('tethysapp.temp_precip_trends.api.datetime')
-    @mock.patch('tethysapp.temp_precip_trends.api.overlap_ts')
-    @mock.patch('tethysapp.temp_precip_trends.api.param_check')
-    @mock.patch('tethysapp.temp_precip_trends.api.get_cum_precip_data')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.log')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.relativedelta')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.datetime')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.overlap_ts')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.param_check')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.get_cum_precip_data')
     def test_get_proj_cum_precip_exceptions(self, mock_get_cpd, mock_param_check, mock_overlap_ts, _, __, ___):
         mock_request = self.factory.get('/get-proj-cum-precip/', {'end_time': 'test'})
 
@@ -183,12 +183,12 @@ class ApiTests(TestCase):
 
         self.assertEqual(expected_result, json.loads(ret.content))
 
-    @mock.patch('tethysapp.temp_precip_trends.api.param_check')
+    @mock.patch('tethysapp.temp_precip_trends.controllers.api.param_check')
     def test_bad_params(self, mock_param_check):
         mock_request_min = self.factory.get('/get-min-temp/')
         mock_request_mean = self.factory.get('/get-mean-temp/')
         mock_request_max = self.factory.get('/get-max-temp/')
-        mock_request_tp = self.factory.get('get-total-precip')
+        mock_request_tp = self.factory.get('/get-total-precip/')
         mock_request_cp = self.factory.get('/get-cum-precip/')
         mock_request_pmt = self.factory.get('/get-proj-mean_temp/')
         mock_request_pcp = self.factory.get('/get-proj-cum-precip/')
